refactor: 외부 API 조회 Controller 분리 #51
